backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
httpx>=0.24.0
python-dotenv>=1.0.0
anthropic>=0.40.0
